Make config.xml a Jinja2 template
Add Jinja2 template support to the Pyramid app and change the config.xml
file into a rendered config.xml.jinja2 template.

This is so variables from the Python code can be used for the URLs in
the config.xml file. In a following commit these variables can now be
read from environment variables.