Lint the tests as well
Test code can also contain mistakes, and should also have consistent
style etc, so run prospector over the Python test code as well.

There's some linter configuration that we want to apply to the tests
only and not to the production code. The best way I could find to do
this was to have `make lint` run prospector twice. The first prospector
run is for the production code only (`test-warnings: false` in
`.prospector.yaml`), the second run is for the test code only.

The second run passes the alternative config file
`tests/prospector.yaml` to prospector on the command line. This config
file inherits from the main `.prospector.yaml` and then adds test-only
config on top.

This isn't perfect because `make test` prints out two separate
prospector reports, but it's the best way I could get it working.