Correct erroneous /token_init URL
Our Canvas app was using /token_callback as the redirect_uri in the
OAuth 2.0 authorization request but then later, in  the access token
request, it was using /token_init as the redirect_uri.

/token_init doesn't exist (404) and the OAuth 2.0 spec says that the
redirect_uri URI params in the two requests must be the same:

https://tools.ietf.org/html/rfc6749#section-4.1.3

So /token_init appears to have been a typo. Correct it to
/token_callback.

The docs also say to enter /token_init as the redirect URI when creating
the developer key in Canvas. Change this to /token_callback as well.