Turn AuthData into AuthDataService
Add the pyramid_services Pyramid extension:

  https://github.com/mmerickel/pyramid_services

and turn AuthData into AuthDataService, a pyramid_services service.

This follows the pyramid_services-based service layer that h has.

This means that auth_data.py moves from lti/models/ into lti/services/
and is accessed as request.find_service(name='auth_data') instead of
as request.auth_data.

Otherwise everything's the same.

AuthDataService still reads and writes the canvas-auth.json file itself.
In the future we'll introduce an SQLAlchemy database and multiple ORM
classes (along the lines of OAuth2Client and OAuth2AccessToken) to
replace the canvas-auth.json file. When that happens AuthDataService will
no longer read and write a canvas-auth.json file, nor will it read and
write to the db itself, rather it'll orchestrate multiple ORM classes
that talk to the db themselves.

At this point AuthDataService will become a true service - something that
sits between the views and the models and orchestrates multiple model
classes on behalf of the views.