Save submitted credentials to database
Save OAuth 2.0 credentials (client ID and secret) submitted to us via
our /lti_credentials form to the database instead of to a
credentials.txt file.

This only supports the Path A version of the /lti_credentials form, but
Path B hasn't actually been removed from the form yet.

A new database table, oauth2_unvalidated_credentials, is added to store
the completely unvalidated credentials submitted via this form. These
submitted credentials will have to be moved, manually, into the
canvas-auth.json file or in future the oauth2_credentials db table in
order to activate the app for a particular set of credentials.