Use Via for PDF assignments
This is just a test of whether the Canvas app can just point Via to the
Canvas file's download URL, rather than downloading the file from Canvas
to its own disk and serving it, having its own vendored copy of PDF.js,
and pointing PDF.js at the downloaded file.

This will certainly be slower, because Via is slow, but we hope to
replace Via with something faster soon anyway.

If this works it simplifies the Canvas app a lot.

I'll need to merge this and push it to QA in order to test it.