Revert "Remove QA and prod deploys from Jenkinsfile"
This reverts commit d01843851b7abc6ce8065b7ffbf2d4aa12de2e98.

I think we've made the necessary changes in the playbook repo, the
deployment repo, and Jenkins now. I'm not sure if these qa and prod
deploys will run successfully in Jenkins yet, I think we may need to
manually create lms-qa and lms-prod environments in Elastic Beanstalk
first, but lets give this a try and find out.