Move alembic.ini file
Move the alembic.ini file into the conf dir. This puts the file in the
same dir as other .ini config files, in the same location as it's in in
all our other repos, means that scripts such as those in
https://github.com/hypothesis/deployment work, and means that most of
the alembic commands given in the README are correct again.