Remove SSL-in-dev support
Remove support and docs for running the app in SSL in dev.

This was done so that your local dev app can be installed in the hosted
https://hypothesis.instructure.com/ without triggering mixed content
warnings from the browser. But it makes development a pain as you
repeatedly have to bypass self-signed certificate warnings, things go
wrong and show empty responses in the browser, etc.

Running your local dev app over SSL also doesn't work well when you want
to use it with your local h, client and via, to avoid mixed content
warnings you would have to run all of those over SSL as well.

It's much simpler to just click the button in Chrome to bypass the mixed
content warning.