Temporary fix for migrations on qa/production
The `from lms import db` in `migrations/env.py` is failing with
`ModuleNotFoundError: No module named 'lms'` on QA and production,
though it works in development environments.

Work around this for now by duplicating the naming convention in
`migrations/env.py` instead of importing it.

We should figure out why the important is failing and fix it properly at
some point.