Add an assert statement
The `group` object that add_user_to_group() retrieves from
CourseGroup.get() is never expected to be None -- as noted in the
assertion comment the group should always have been created before this
function is called.

If this assumption is violated and `group` _is_ None then an
AttributeError is raised later on in the code when an attempt is made to
access group.pubid.

This commit changes this AttributeError to an AssertionError that's
raised immediately after the unexpected `group` value is received and
has a helpful, human-written error message attached.

Additionally, the implicit assumption that `group` is never None is now
explicit in the code.