Move JavaScript for content_item_selection view out of template
Move the script for the form used to select a document to annotate out
of the Jinja2 template and into a JS module. This will make it easier to
refactor in future.

This commit makes only a minimum of changes to the actual code itself
needed to support loading the script from a separate file instead of
being inline.

 - The functions invoked via event handlers in the template are exposed
   via a "contentItemSelection" window global.
 - All template-variables have been replaced with access to properties
   on the `window.DEFAULT_SETTINGS` global.
 - Remove unused `appId` variable.
 - Fix lint failures in the code