Add js-hypothesis-config to LTI launch HTML
Add a new variable `{{ context.hypothesis_config }}` to the context for
all templates.

This variable contains a js-hypothesis-config JSON object, suitable for
rendering in a `<script type="application/json" class="js-hypothesis-config">`
tag in order to configure the client. See:

https://h.readthedocs.io/projects/client/en/latest/publishers/config/#configuring-the-client-using-json

Use this variable in the new_lti_launch.html.jinja2 template to render a
js-hypothesis-config JSON script.

Nothing uses this js-hypothesis-config JSON script yet. A Via iframe is
injected into the LTI launch response, but the Hypothesis client itself
is _not_ injected in the LTI launch HTML response itself, so it will not
read this js-hypothesis-config script. The script needs to be read and
passed to the client by something else, which will be done in a future
commit.

The changes in this commit require two new config settings that're
needed to generate the `grantToken` part of the client config. Two new
environment variables are added to the install instructions:

* H_JWT_CLIENT_ID
* H_JWT_CLIENT_SECRET