Rewrite "Use Google Picker" button handler to avoid popup blocker
Rewrite the "Use Google Picker" button handler to use promises to manage
async setup of the Google API client libraries instead of a series of
interdependent callbacks which were hard to reason about.

This made it possible to initialize the Google API client libraries and
invoke `gapi.auth2.init` as soon as the API entry script loads. This in
turn avoids the popup blocker showing when `gapi.auth2.signIn` is called
later on.

 - Remove `enablePickerButton` function which didn't actually do
   anything.
 - Use `gapi.auth2` instead of `gapi.auth`. The latter is not mentioned
   in the current Google client library documentation and presumably
   deprecated.
 - Remove state variables that are not required after switching to use
   promises.

Fixes #186