Add a JavaScript JSON-RPC server
Add a JSON-RPC server that responds to JSON messages sent over
postMessage.

Currently the only request that is supported is "requestConfig" which
returns a client config object for the current LTI launch request. This
client config object was already rendered into the HTML document by a
previous commit, and the JSON-RPC server reads it from there.

The JSON-RPC server's own config object, including the list of
postMessage origins from which RPC requests will be accepted, was also
rendered into the document by a previous commit and is now read by the
JSON-RPC server.