Add js-rpc-server-config to template
Add a JSON configuration object to the LTI launch HTML response for the
(not yet implemented) JavaScript JSON-RPC-over-postMessage server that
is to be added in a future commit. The config object contains the list
of origins from which the server will accept RPC requests. The server
will find the the config object by its CSS class when it starts up.

An additional required config setting / envvar, `RPC_ALLOWED_ORIGINS`,
is added to support this new setting.