Deduplicate LTI param retrieval
Deduplicate some code for retrieving stashed LTI launch parameters from
the database that was duplicated in a couple of places.

This code is going to be needed in a third place too, in an upcoming
commit.

The `find_by_state()` function is no longer used outside of the
oauth_state.py file itself, so make it a private function.