Fix Bad Request when rendering Canvas file assignments
When rendering a Canvas file assignments the request is actually an
OAuth 2.0 redirect request from Canvas, not an LTI launch request, so
request.params is the OAuth params not the launch params. `Root` was
trying to access launch request params from `request.params` and raising
Bad Request when it didn't find them. The launch params have been
stashed in the database during a previous LTI launch request and need to
be retrieved from there. Fix `Root` to detect when it needs to retrieve
the launch params from the DB and do so.