Fix exception handling using pyramid_exclogger
Fix the in-browser exception views (error page) and printing of
exceptions to the console in dev.

This was all just completely broken. Fix it by using pyramid_exclogger
to log exceptions to the console, and also fix up the exception views /
error page to not crash, and tidy up the code and docstrings.