Use debugtoolbar instead of exclog in dev
Use pyramid_debugtoolbar instead of pyramid_exclog to handle exceptions
in dev. When done correctly debugtoolbar will show the traceback
(rendered nicely) inline in the browser, give you a link to the
/_debug_toolbar/ page with lots more info, _and_ print the traceback to
the terminal.

In production still use pyramid_exclog and do not use debugtoolbar.

So in dev we use debugtoolbar only and in production we use exclog only.

To allow pyramid_debugtoolbar to print tracebacks to the terminal in dev
we need to **not** register the app's exception views in dev.
debugtoolbar doesn't print tracebacks from "squashed" exceptions to the
terminal (though it still does give you access to it all in the
browser). "Squashed" means the exception was caught by an exception
view. Since we have an `@exception_view_config(Exception)`, that is, we
catch _all_ exceptions and always show the user a nice error page, all
exceptions in our app are squashed exceptions and none are ever printed
to the terminal by debugtoolbar.

So change the registering of the error views to be imperative rather
than declarative, and only register the error views if the "debug"
setting is not set. This setting is set in development.ini.

This means that:

1. In production the error views will be registered and the nice error
   page will be shown (and errors will be reported to Sentry)

2. In development the error views will not be registered. On an
   exception pyramid_debugtoolbar will render the exception in the browser,
   the /_debug_toolbar/ page will work, and the exception will be printed
   to the terminal.

3. In development if you set the envvar `export DEBUG=false` then you
   will get the error pages as in prod, so you can test those.

4. Our actual error views do not need to contain any "if debug then X
   else Y" logic, such as re-raising exceptions for printing to the
   terminal in dev.