Don't strip launch params from content item form
Sometimes when the app receives an LTI launch request it needs to do
something else before launching an assignment:

* In LMS's that don't support LTI content item selection the URL of the
document to be annotated will not have been chosen during the assignment
creation process, and it instead needs to be chosen the first time the
assignment is launched before doing the actual assignment launch.

* When a Canvas file is being annotated the user needs to be sent to
Canvas's OAuth authorization page first, before launching the
assignment.

In either case when the app receives a launch request it needs to send
the user to another page to do something first, and _then_ it needs to
do the originally requested launch request. For example show the user a
form for entering the URL of the document to be annotated, the user
pastes in a URL and clicks Submit and _then_ the app does an LTI launch
using the originally received launch params.

The app implements this re-doing of an LTI launch by actually containing
its own LTI launch form (new_content_item_selection.html.jinja2) which
it fills with the original launch params and submits to its own launch
URL.

content_item_form() is the helper function that returns the launch
params that're passed to the new_content_item_selection.html.jinja2
(launch form) template. Because this function uses a hardcoded whitelist
of launch param names, all of the other params from the original launch
request, some of which we're using, get lost.

This is one of the reasons why the "provisioning" features aren't
working in some LMS's, such as Blackboard.