Add a Hypothesis API service
The pyramid_services package was already registered as a dependency in
requirements.in but the extension was not being included in our app. Add
config.include("pyramid_services") to lms/app.py to enable the
extension.

Also add a lms/services package to contain our services and include that
in app.py too.

Add a HypothesisAPIService class to lms/services. This object is
accessable as request.find_service(name="hapi") and has low-level
methods for making Hypothesis API requests with error handling.

Move the post() wrapper function into the new
HypothesisAPIService where it becomes HypothesisAPIService.post().
Change view decorator code to use the new HypothesisAPIService instead.