Fix the styling of the reports page
The styling of the reports page was such that, once there's enough data,
the newest table rows are in a fixed and out-of-view position and can't
be seen.

Change application_report.html.jinja2 to be a standalone template and no
longer extend base.html.jinja2. This means that the global lms.css file
no longer applies to this page. Add a reports.css file for the page
instead containing some minimal CSS to make the tables look slightly
nicer.