Make new_lti_launch.html.jinja2 inherit from base
Make new_lti_launch.html.jinja2 inherit from the base template. This
means that the contents of our iframe is now a complete HTML document,
rather than some HTML with no <html>, <head> or <body> as before.

It also means that stuff added to the base template in the future, such
as JSON config objects, will now be present in the
new_lti_launch.html.jinja2 page as well.