Move authentication policy into lms.authentication
I'm going to be adding more authentication stuff and I need this to be
here, not in config which is totally the wrong place for it.

Note that the authorization policy is still being set in lms.config.
That should move too, to lms.authorization, but I don't need to do that
right now.