Add a schema for validating LTI launch params
Add `lms.validation.LaunchParamsSchema` -- a marshmallow schema for
validating LTI launch params. It includes a
`LaunchParamsSchema(request).lti_user()` method that returns the
`LTIUser` object for the request if it's a valid LTI launch request.

Not all LTI launch params are validated yet -- only the ones that we
need right now.