Add lms.values.LTIUser value type
Add a "value object" type for an LTI user -- an immutable object with no
methods that just exposes a set of user-related fields: user_id,
oauth_consumer_key (more can be added in future).

In the future an instance of this class will be exposed as
`request.user` whenever an LTI user is authenticated, to contain more
info about the user than just `request.authenticated_userid`.