Add an authentication policy for LTI users
Add a Pyramid authentication policy that sets
`request.authenticated_userid` to a userid string based on the LTI user
whenever an LTI user is authenticated either via launch params or a
bearer token.

This is in addition to the existing cookie-based
AuthTktAuthenticationPolicy that's used for the `/reports` page. If no
launch params or bearer token are found in the request then the new
authentication policy delegates to the old AuthTktAuthenticationPolicy.