Add authenticated_user() helper function
Add a helper function for generating a Pyramid
`request.authenticated_userid` string from an LTIUser value.

This requires us to decide what a `request.authenticated_userid` string
should look like for an LTI user. I've gone with:

    <USER_ID>:<OAUTH_CONSUMER_KEY>

where `<USER_ID>` is the value of the LTI launch param `user_id` (i.e.
the user ID from the LMS) base 64 encoded so that we know it doesn't
contain `":"`. And `<OAUTH_CONSUMER_KEY>` is the value of the
`oauth_consumer_key` launch param which we know doesn't contain `":"`
because it was originally generated by us (and the given
oauth_consumer_key has been verified to match one in our DB before
setting `request.authenticated_userid`).