Pin modal dialogs to the top of the page
Pin modal dialogs to the top of the page (with a margin of 80px from the
top of the page) instead of vertically centering them.

This prevents dialogs from jumping around when the height of a dialog's
contents changes (e.g. a new element gets added to the bottom of the
dialog) and the browser vertically re-centers the dialog using its new
height.