Add /flags page for toggling feature flags
Add a new required setting `FEATURE_FLAGS_COOKIE_SECRET` /
`feature_flags_cookie_secret`, which is a secret used to sign a JWT that
is the value of the feature flags cookie.

Also add an optional `FEATURE_FLAGS_ALLOWED_IN_COOKIE` /
`feature_flags_allowed_in_cookie` setting that is the list of feature
flag names that're allowed to toggled on or off in the cookie and that
will be shown on the `/flags` page.

Then add the `/flags` page itself, that lets you read and write the
feature flag toggles in your cookie. This is just the page for reading
and writing the cookie, but the cookie isn't actually used to determine
feature flag state yet.

The `/flags` page has its own Jinja2 template and some CSS of its own in
`feature-flags.css` but mostly it inherits the base template and CSS
from `base.html.jinja2`. This couples the feature flags extension, which
is meant to be a separable extension, to the app's base template and
CSS. When the feature flags extension is extracted into its own repo
this base template and CSS will have to be extracted (and duplicated)
with it.