Add feature flag view predicate
Add a `feature_flag="my_feature"` view predicate (for use in Pyramid
`@view_config(...)`'s) that can be used to toggle a view on or off
depending on a feature flag, or to direct requests to one view or
another depending on a feature flag.

Usage:

    @view_config(route_name="foo", feature_flag="foo")
    def view_that_will_be_called_when_feature_flag_is_on(request):
        ...

    @view_config(route_name="foo")
    def view_that_will_be_called_when_feature_flag_is_off(self):
        ...

Also add a test page at `/flags/test/view-predicate`.