Don't show login form other than for /reports
Don't show the login form (https://lms.hypothes.is/login) except when
the user is trying to visit the reports page
(https://lms.hypothes.is/reports) and is not logged in.

The type of login provided by the `/login` form is only used by the
`/reports` page. If an unauthenticated user tries to visit some other
view that requires authentication (for example a view that requires
authentication by LTI launch params, or by a bearer token, maybe an API
view) it makes no sense to respond with the login form.

Unauthenticated requests to these other authentication-requiring views
will now use Pyramid's default forbidden view (a 403 Forbidden response
whose body is a plain HTML error page) instead.