Add Canvas API authorize endpoint
Add a new `/api/canvas/authorize` endpoint, intended to be opened by the
frontend in a popup window when we need the user to authorize us to
access their Canvas account. The server's response to a
`/api/canvas/authorize` is an HTTP redirect to the user's Canvas
instance's authorize page, so what the user sees in the popup window is
the Canvas instance's authorize page.

Opening this popup window is the first step in the OAuth 2 authorization
code grant flow.

Also add a new stub `/canvas_oauth_callback` page, used only when the
`"new_oauth"` feature flag is on, that handles Canvas's redirect from
its authorize page back to us. This page has yet to be filled with the
actual implementation.

This popup window endpoint is still missing a bunch of functionality,
including:

- [ ] Storing and verifying the OAuth 2 state param, crucial for security
- [ ] Handling unauthenticated `/api/canvas/authorize` requests
- [ ] Robust handling of the unvalidated and unverified LMS URLs in our DB
- [ ] Handling of application instances for which we have no developer key