Refactor Jenkinsfile to run things in parallel
Use Jenkins Pipeline's `parallel` to run everything in separate
paraellel stages, rather than using `make -j` to do the same.

There were intermittent build failures with `make -j` due to a (probably
fixable) race condition with tox. `make -j` also didn't enable the
frontend tests to be run in parallel with everything else because the
frontend tests need to be run from a different Docker image (and
therefore a different instance of `make`. Finally, with `make -j`
Jenkins is unaware of the concurrency whereas with Jenkins Pipeline
parallel stages Jenkins can show the different stages running
concurrently in its UI and can label the stdout and stderr with which
concurrent stage it comes from.

The new `Jenkinsfile` first runs a `"Build"` stage and then runs
`"Formatting"`, `"Docstrings"`, `"Backend lint"`, `"Backend tests"`,
`"Frontend lint"` and `"Frontend tests"` stages in parallel.

As before, all the stages are run in a Docker container based off the
production Docker image but with additional test dependencies installed.

Also as before the frontend tests use a separate Node Docker image due
to compatibility problems between the frontend tests and the production
Docker image.

The Postgres "sidecar" container is run only for the `"Backend tests"`
stage.

Some tidying up and deduplicating of the `Jenkinsfile` has been done.

`make checkdocstrings` has been added - this wasn't being run on CI
before.

Limitations
-----------

The main limitation of all this is that test dependencies --
dependencies that are not part of the production Docker image but that
need to be installed in the Docker containers to run the various test,
lint, formatting and docstring commands on Jenkins -- are being
installed multiple times over, once for each stage `"Formatting"`,
`"Docstrings"`, `"Backend lint"`, `"Backend tests"`, `"Frontend lint"`
and `"Frontend tests"`. Each stage's dependencies are not exactly the
same, but there's a lot of overlap. A way to avoid doing the same work
multiple times would be good.

I'm not sure that this is a huge limitation though -- I'm not sure that
there's much build time to be shaved off by avoiding duplicate
dependency installs. It may not be worth the complexity it would
introduce. And build times are already pretty good (2-3mins) anyway.

sitepackages
------------

This commit also enables tox sitepackages on Jenkins.

So that it reuses the installed sitepackages (i.e. the Python packages
already installed in Alpine Linux) instead of reinstalling them all in
tox's venv.

This seems to speed up Jenkins builds by about 20-30 seconds in total.