Require LTI authentication for Canvas API popup
`/api/canvas/authorize` is the popup window that frontend code opens
when it wants to start an OAuth 2 authorization flow with the Canvas
API. Require the `"canvas_api"` permission for this view so that, if the
URL is requested without LTI authentication, the view won't be called
and an authorization error (403) will be sent back.

When launching the app within Canvas frontend code will use one of our
LTI JWTs in a query param when opening `/api/canvas/authorize` so the
permission check will succeed.

But if you try to visit `/api/canvas/authorize` in a tab directly you'll
get an authorization error page.