Remove schema from canvas_oauth_callback
Remove CanvasOAuthCallbackSchema from the canvas_oauth_callback() view.

This fixes an issue that, since the frontend code without the new_oauth
feature flag doesn't stash the new OAuth 2 state param in the session
cookie, it was getting a validation error and the app was effectively
broken.

This means that canvas_oauth_callback() can now be called with invalid
params and will crash, which is how it was before this schema was added
to the view.

The new_oauth work is going to supercede this view entirely anyway.