Support reading LTI user from OAuth2 state param
Extend `CanvasOAuthCallbackSchema` to support reading the LTI user from
an OAuth 2 `state` query parameter.

`CanvasOAuthCallbackSchema.state_param()` is a method that generates and
returns a `state` param capable of authenticating the LTI user. The
`state` param is a signed and timestamped JWT containing the current
`request.lti_user` and a CSRF token that's also added to the session. A
view should call `state_param()` to get the value for the
`?state=<VALUE>` parameter for putting into a Canvas authorization
endpoint URL before redirecting the browser to it.

`CanvasOAuthCallbackSchema.lti_user()` is a method that returns the
`LTIUser` value from the current request's `state` param, after
validating the JWT. An authentication policy should call to authenticate
`redirect_uri` requests.

The JWT and CSRF token are also verified before calling the view, when
`CanvasOAuthCallbackSchema` is used as a schema to guard a view. The
CSRF token is also removed from the session so that it can't be reused.