Exchange authorization codes for access tokens
After receiving authorization from the user (via the popup window)
exchange the authorization code for an access token in a server-server
request to the Canvas API.

The access token and its refresh token and expiry time get saved to the
DB but otherwise they aren't used yet.

If the DB already contains an access token for this user (same consumer
key and user_id) that token will be updated rather than saving a second
one.