Read authorization param from form fields
Support reading authorization JWTs from form fields, as well as from
headers and query strings.

We'll need this in future as we're going to have a need for
authenticating form submissions (and using a form field seems nicer than
a query string on the form's `action` URL).

I didn't add a test for this because I couldn't find a way to create a
test that would fail if only "querystring" (and not "form") was in the
`locations=[...]` list and then would pass when "form" was added. Adding
the `authorization` param to the dummy request's `.params` will work
with either "querystring" or "form" or both, for example.