Add report_lti_launch view decorator
This can be added to LTI launch views like so:

    @view_config(
        decorator="lms.views.decorators.report_lti_launch",
        ...
    )
    def my_lti_launch_view(context, request):

For details see the decorator parameter documented here:

https://docs.pylonsproject.org/projects/pyramid/en/latest/narr/viewconfig.html?highlight=decorator#non-predicate-arguments

It will add a row to the `lti_launches` DB table each time the view is
called. Rows from this table are the data used by the `/reports` page.

New views will use this instead of a helper function that the legacy
views used to append to the same table:

https://github.com/hypothesis/lms/blob/d06cae8bb0778867588c3f7e95c2050385a8776d/lms/views/lti_launches.py#L80-L93

Nothing uses the new view decorator yet.