Add unique constraint to module_item_configurations
In practice no two module_item_configurations with the same
resource_link_id and tool_consumer_instance_guid are ever
created, and if a duplicate did get into the DB the code would
misbehave.

See:

https://github.com/hypothesis/lms/blob/3430effa15d8f8137db908f5c40ab1df5c85f908/lms/views/lti_launches.py#L70-L77

So add a DB constraint to make sure that this can't happen, and also to
speed up a common DB query by adding an index.