Add ConfigureModuleItemSchema
This schema will be used to protect the `/module_item_configurations`
route. When an unconfigured assignment is launched the content item form
that's rendered gets submitted to `/module_item_configurations` so that
the chosen document URL can be saved in the DB. This schema will be used
to validate those form submissions.