Add LTIUser.roles value
LTIUser objects are the values that get serialized into JWT
authorization params that are then included in XHR requests, form
submissions and OAuth 2 redirects, deserialized again, and used to
authenticate the user making the request.

We have an upcoming need to know the user's LTI roles as well as their
user_id and oauth_consumer_key, so add a new field LTIUser.roles.