Don't crash on unauthorized LTI launches
Even though the views that the `authorized_to_configure_assignments`
predicate is used with also have the
`permission="launch_lti_assignment"` predicate, so the view won't
be called if the request isn't an authenticated LTI launch request, that
doesn't mean that the `authorized_to_configure_assignments` predicate
won't be called for requests that don't have permission. That depends on
whether Pyramid evaluates the `authorized_to_configure_assignments` or
the `permission` predicate first, and I'm not sure that the order in
which Pyramid evaluates predicates is defined.

This means that the `authorized_to_configure_assignments` predicate
might be called when `request.lti_user` is `None` and would crash. Fix
it to say that the user isn't authorized in this situation.

This also means that `authorized_to_configure_assignments` can safely be
used on views that don't have `permission="launch_lti_assignment"` and
it'll just say the the request isn't authorized.