Make the Google deployment settings required
There are only two deployments of this app - lms.hypothes.is and
qa-lms.hypothes.is - and both of those have the Google Drive-related
environment variables set. This app isn't intended to be used by third
parties. So there's no need for the Google Drive-related settings to be
optional, and the code can be simpler if it assumes that these settings
are always present.

In development environments, if the developer wants to test the app
without going to the trouble of creating a Google Drive client, they can
just set these envvars to placeholder values as the README already says
to do. Everything will work fine except that clicking on the
**Select PDF from Google Drive** won't work.

We could make this even easier for developers by providing placeholder
values in `development.ini` but that isn't done for any of the other
settings yet.