Move Python requirements into requirements files
Move all Python requirements out of `tox.ini` and into
`requirements-*.txt` files, and pin them to exact versions. This means
that:

1. The exact versions of dependencies that get installed won't vary from
   one developer to the next, between dev and CI, or over time, except
   when we update the `requirements-*.txt` files.

2. Dependabot can send us PRs to update the `requirements-*.txt` files.
   This is the reason for moving them into requirements files --
   Dependabot doesn't support updating requirements listed in `tox.ini`
   files.

The first step was to move all our lists of Python requirements --
requirements for linting, requirements for running the tests, etc -- out
of `tox.ini` and into `requirements-*.in` files. Then compile each
`requirements-*.in` file to a corresponding `requirements-*.txt` file
using `pip-compile`.

This results in a large number of different `requirements-*` files. In
the future we may want to merge some of these together to reduce the
number of them.