Don't fetch Canvas file URLs on launch
Fix `canvas_file_basic_lti_launch()` to not fetch the Canvas file's
public URL from the Canvas API as part of the initial response to the
launch request. Instead add the proxy API URL that will return the
Canvas file's URL to the `"urls"` in the JavaScript config object.

The idea is that JavaScript code will make an XHR request to the
`via_url` proxy API and use the Canvas file URL from the response to
inject the Via iframe into the DOM.

This means that the server's response to the initial launch request can
be fast as it doesn't need to talk the the Canvas API, and all requests
to the Canvas API are kept within the `/api/canvas/` proxy API views.

For more reasoning see
https://github.com/hypothesis/lms/issues/622#issuecomment-499909135