Add error handling to OAuth 2 redirect view
When we send the access token request to Canvas, catch any error
responses or network timeouts etc and show the user an error page with a
"Try again" button.

`CanvasAPIClient.get_token()` now raises `CanvasAPIServerError` if
there's a problem with the access token request.

The `oauth2_redirect()` view that calls `CanvasAPIClient.get_token()`
catches the `CanvasAPIServerError` and wraps it in an
`HTTPInternalServerError`.

This causes Pyramid to call the new `oauth2_redirect_error()` exception
view which renders a template with a "Try again" button that links back
to the original `authorize()` view to restart the flow.