Get rid of `instantiate_schema()`
`instantiate_schema()` was a private helper of `lms.validation` that was
used to instantiate our schemas because we supported schemas being
either webargs schema dicts or marshmallow schema classes.
`instantiate_schema()` would detect whether the schema was a class and
if so instantiate it.

In practice all our schemas are now marshmallow classes as these are
more powerful while still being compatible with webargs, and really no
harder to write than webargs schema dicts. Having all schemas be the
same is consistent.

So get rid of `instantiate_schema()`, which is no longer necessary, and
just instantiate the schema classes directly.